Give NbnxmGpu a constructor
Renamed variable without underscore for consistency

Refs #5539

+++ src/gromacs/nbnxm/gpu_data_mgmt.h
@@ -44,6 +44,7 @@
 #define GMX_NBNXN_GPU_DATA_MGMT_H
 
 #include <memory>
+#include <optional>
 
 #include "gromacs/gpu_utils/gpu_macros.h"
 #include "gromacs/mdtypes/interaction_const.h"
@@ -91,8 +92,8 @@ NbnxmGpu* gpu_init(const DeviceStreamManager gmx_unused& deviceStreamManager,
                    const PairlistParams gmx_unused&      listParams,
                    const nbnxn_atomdata_t gmx_unused*    nbat,
                    /* true if both local and non-local are done on GPU */
-                   bool gmx_unused                     bLocalAndNonlocal,
-                   const std::optional<int> gmx_unused nLambda) GPU_FUNC_TERM_WITH_RETURN(nullptr);
+                   bool gmx_unused bLocalAndNonlocal,
+                   const std::optional<size_t> gmx_unused nLambda) GPU_FUNC_TERM_WITH_RETURN(nullptr);
 
 /** Initializes pair-list data for GPU, called at every pair search step. */
 GPU_FUNC_QUALIFIER
